Fixed some logic with group membership, namely: - Added users to the member of groups and organizations. - Added workflows and tools as members.
Also, changed user creation code to ensure that CKAN users are created
with the same UUID as Keycloak users. This will allow us to directly
identify the two (instead of the more complex and error-prone practice
of having just the username match). This was done as part of the new
code but will need to be tested more thoroughly.